Added respond_to? extension to wrapper class
When respond_to? is called on a instance of the wrapper class, it would
technically give an incorrect false when a method could just be passed
along to the connection object via method_missing. This resolves that
issue by checking the instance methods of the connection object along
with the instance methods of the wrapper when respond_to? is called.