Checking in pushes objects to the head of the queue.
This means that, when not saturated, objects in the pool will be reused
more often.

This behavior allows you to create a pool of objects which in turn have
some kind of time outs themselves. For instance, objects which keep an
underlying TCP connection open until they're idle for some time. An
example of this is `Net::HTTP::Persistent`.

See also issue #20.