Reclaim old connections in TimedStack
This adds an LRU to TimedStack so that the oldest connection type will
be destroyed so a new type can be created.

This only works on ruby 1.9+ because Hash is ordered.  I'm unsure if
this is OK as the tests do not run on Ruby 1.8 (missing
require_relative).  I'm guessing this means that connection_pool does
not support 1.8 even if that is not officially stated anywhere.