Ability to optioanlly drop all connections after fork
There was a recent feature to automatically drop all connections after fork. This is quite nice
and makes sense.

However, for some rails app that usually follow the fork model (like w/ unicorn/puma), and additionally
have some logic to fork processes to perform internal business logic that doesn't rely or use ConnectionPool,
the application can observe Redis connection issues or resets. These forks can happen during application run time.
Like ours.

In such a case, it'd be nice to not automatically drop all the connections, since the underlying process isn't working
with Redis/ConnectionPool, and as a sideeffect the pool in the primary process is impacted.

This PR proposes a new attribute auto_reload_after_fork as a config option. By default it is true. However, application
users can turn it to false and not opt in for the feature to auto drop connections after fork.

This could be quite useful for us